requirements.txt
streamlit==1.44.1
pypdf==4.2.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
python-dotenv==1.0.1
groq==0.23.1
nltk==3.8.1